requirements.txt
anyio==3.7.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.2
cssselect==1.2.0
cssutils==2.7.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
emails==0.6
fastapi==0.98.0
Flask==2.3.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka==1.3.5
kafka-python==2.0.2
lxml==4.9.3
MarkupSafe==2.1.3
passlib==1.7.4
premailer==3.10.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
secure-smtplib==0.1.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.6
